+++ src/main/java/pe/edu/upc/avi_aplicaction/repositories/IRolesRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Repository
 
-public interface IRolesRepository extends JpaRepository <roles, Integer> {
+public interface IRolesRepository extends JpaRepository <roles, Long> {
     @Query("SELECT r FROM roles r WHERE r.rol LIKE %:Tipo%")
     public List<roles> buscarRol(@Param("Tipo") String Tipo);
 }
