+++ src/main/java/pe/edu/upc/avi_aplicaction/serviceimplements/UsersServiceImplements.java
@@ -28,12 +28,12 @@ public class UsersServiceImplements implements IUsersService {
     }
 
     @Override
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         uR.deleteById(id);
     }
 
     @Override
-    public Users listUserById(int id) {
+    public Users listUserById(Long id) {
         return uR.findById(id).orElse(new Users());
     }
