+++ src/main/java/pe/edu/upc/avi_aplicaction/serviceinterfaces/IRolesService.java
@@ -7,7 +7,7 @@ public interface IRolesService {
     public List<roles> list();
     public void insert(roles role);
     public void update(roles role);
-    public void delete(int id_Rol);
-    public roles listId(int id_Rol);
+    public void delete(Long id_Rol);
+    public roles listId(Long id_Rol);
     public List<roles> buscar(String rol);
 }
